Step 2: Extract rhythm classes into rhythm.py
Move beat, rest_symbol, continue_symbol, rest, and rhythm classes
into new musicpy/rhythm.py module. These classes are imported back
into structures.py to maintain backward compatibility.

Uses lazy mp import and lazy parsing imports to avoid circular dependency.
